moved buy book from toolbar to bottom of page, changed the styles, added common css for buttons in toolbar, updated snaps
