+++ src/app/content/components/Toolbar/HighlightButton.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
-import styled, { css } from 'styled-components/macro';
+import styled from 'styled-components/macro';
 import highlightIcon from '../../../../assets/highlightIcon.svg';
 import { useAnalyticsEvent } from '../../../../helpers/analytics';
-import theme from '../../../theme';
 import { AppState, Dispatch } from '../../../types';
 import { openMyHighlights as openMyHighlightsAction } from '../../highlights/actions';
 import * as selectors from '../../highlights/selectors';
 import { toolbarIconStyles } from './iconStyles';
-import { PlainButton, toolbarDefaultText } from './styled';
+import { PlainButton, toolbarDefaultButton, toolbarDefaultText } from './styled';
 
 interface Props {
   openMyHighlights: () => void;
@@ -18,14 +17,8 @@ interface Props {
 
 // tslint:disable-next-line:variable-name
 const MyHighlightsWrapper = styled(PlainButton)`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 2rem;
+  ${toolbarDefaultButton}
   height: auto;
-  ${theme.breakpoints.mobile(css`
-    margin-right: 0;
-  `)}
 `;
 
 // tslint:disable-next-line:variable-name
