+++ src/app/messages/en/messages.json
@@ -108,7 +108,8 @@
   "i18n:toolbar:highlights:popup:body:use-this-page": "Then use this page to create your own study guide",
   "i18n:toolbar:highlights:popup:help-info": "Need help? Visit the <a data-testid='support-link' data-analytics-label='support-link' target=\"_blank\" href=\"https://openstax.secure.force.com/help/articles/FAQ/Using-the-highlighting-and-note-taking-feature-on-mobile-or-tablet?search=highlighting\">Support Center</a>",
   "i18n:toolbar:highlights:popup:button:back-to-top": "Go back to top",
-  "i18n:toolbar:buy-book:text": "Buy book",
+  "i18n:toolbar:buy-book:aria-label:text": "Buy book",
+  "i18n:toolbar:buy-book:text": "Order a print copy",
   "i18n:cta:heading": "Sign up and continue reading. It's 100% free.",
   "i18n:cta:text": "Read, search, highlight, add notes, build study guides -- all for <strong>free</strong>.",
   "i18n:cta:button:signup": "Sign up",
